What's up, it's me, bruce, And have I got an update for you. Deets in description
Yeah there's the player changes and that wacky camera man, uhh it kinda works, kinda doesn't. Doesn't update for the player going down yet because it's being a bastard but also for going up it's a little bit shit if I'm honest. The grand plan I had is actually wank in practice so maybe we'll have to fiddle around with it a bit to make it good.

Oh yeah there's the interface and shit now as well.